// projects/dynamic-reactive-forms/src/lib/enums/input-type.enum.ts
export enum InputTypeEnum {
  // Checkbox = "CHECKBOX",
  Date = "DATE",
  Email = "EMAIL",
  Hidden = "HIDDEN",
  Number = "NUMBER",
  Password = "PASSWORD",
  // Radio = "RADIO",
  // Range = "RANGE",
  Tel = "TEL",
  Text = "TEXT",
  Time = "TIME",
  Color = "COLOR",
  Month = "MONTH",
  Search = "SEARCH",
  // DropDownList = "DROPDOWN",
  Textbox = "TEXTBOX",
  Url = "URL",
  Week = "WEEK",
  DatetimeLocal = "DATETIME-LOCAL",
}